fix: study 도메인 DB 사양에 맞게 변경

+++ src/main/java/katecam/luvicookie/ditto/domain/study/api/StudyController.java
@@ -28,7 +28,7 @@ public class StudyController {
     }
 
     @GetMapping("/{studyId}")
-    public ResponseEntity<StudyResponse> getStudy(@PathVariable Long studyId) {
+    public ResponseEntity<StudyResponse> getStudy(@PathVariable Integer studyId) {
         return ResponseEntity.ok(studyService.getStudy(studyId));
     }
 
@@ -42,7 +42,7 @@ public class StudyController {
 
     // 회원 로그인 검증 필요
     @DeleteMapping("/{studyId}")
-    public ResponseEntity<Void> deleteStudy(@PathVariable Long studyId) {
+    public ResponseEntity<Void> deleteStudy(@PathVariable Integer studyId) {
         studyService.delete(studyId);
         return ResponseEntity.noContent()
                 .build();
