+++ src/main/java/katecam/luvicookie/ditto/domain/study/dao/StudyRepository.java
@@ -3,5 +3,5 @@ package katecam.luvicookie.ditto.domain.study.dao;
 import katecam.luvicookie.ditto.domain.study.domain.Study;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudyRepository extends JpaRepository<Study, Long> {
+public interface StudyRepository extends JpaRepository<Study, Integer> {
 }
